bin/system/jira/TestGenie-BE/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Clients
httpx==0.25.2

# Authentication & Security
PyJWT==2.8.0

# Configuration
python-dotenv==1.0.0

# File Handling

# CORS
